Refactor worker activity helpers
